More collision and pause functionality. Added Timers
Added the timer class which can be added to a room to call a function after a certain amount of time has passed. Made the pause variable in the room to now stop the physics simulation and stop timer as well. Added the ability to specify a certain tag when  checking collision in the collision world, and added functions to get collisions at a certain point.