Added Input manager removed ldtk Entites class
Added the input manager to help catch inputs between fixed update frames. Removed the previous LdtkEntites system of creating prefabs as functions that return an array of components. The new system is a lot more free and less cluttered.  Also changed the method for the pointInPolygon to work with concave polygons.